Cleanup
Formatting changes

Fix launch exception and update docs
